pyproject.toml
[project]
name = "baris"
version = "0.0.1"
description = "2-player online-multiplayer remake of Buzz Aldrin's Race Into Space"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=1.0"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["baris*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
